+++ crates/biome_js_analyze/src/lint/nursery/no_nodejs_modules.rs
@@ -41,7 +41,11 @@ impl Rule for NoNodejsModules {
     type Options = ();
 
     fn run(ctx: &RuleContext<Self>) -> Self::Signals {
-        let module_name = ctx.query().module_name_token()?;
+        let node = ctx.query();
+        if node.is_in_ts_module_declaration() {
+            return None;
+        }
+        let module_name = node.module_name_token()?;
         is_node_builtin_module(&inner_string_text(&module_name))
             .then_some(module_name.text_trimmed_range())
     }