+++ crates/biome_js_analyze/src/lint/nursery/no_undeclared_dependencies.rs
@@ -1,8 +1,8 @@
 use crate::services::manifest::Manifest;
 use biome_analyze::{context::RuleContext, declare_rule, Rule, RuleDiagnostic};
 use biome_console::markup;
-use biome_js_syntax::{AnyJsImportSpecifierLike, TsExternalModuleDeclaration};
-use biome_rowan::{AstNode, SyntaxNodeOptionExt};
+use biome_js_syntax::AnyJsImportSpecifierLike;
+use biome_rowan::AstNode;
 
 declare_rule! {
     /// Disallow the use of dependencies that aren't specified in the `package.json`.
@@ -44,13 +44,8 @@ impl Rule for NoUndeclaredDependencies {
 
     fn run(ctx: &RuleContext<Self>) -> Self::Signals {
         let node = ctx.query();
-
-        if let Some(parent_syntax_kind) = node.syntax().parent().kind() {
-            // Ignore module declaration statements:
-            // declare module "jest";
-            if TsExternalModuleDeclaration::can_cast(parent_syntax_kind) {
-                return None;
-            }
+        if node.is_in_ts_module_declaration() {
+            return None;
         }
 
         let token_text = node.inner_string_text()?;